+++ CHANGELOG.md
@@ -7,11 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support for `displayAsCurrency` prop in calculation element
+
 ## [0.2.4] - 2021-04-27
 
 ### Added
 
-- Support for `displayAsCurrency` prop in calculation element
 - `zoom` property to location form elements submission data
 
 ### Fixed